refactor(win32): added bindings.py, structs.py and constants.py Those files should be used to access win32 APIs through ctypes. They hold "python-friendly typed" wrappers around win32 APIs. refactor(systray): moved hicon_to_image to app_icons.py moved win_wrappers to bindings.py moved win_types to structs.py moved win_constants to constants.py
